+++ refract/namespace.py
@@ -5,3 +5,7 @@ from refract.elements import (Number, String, Boolean, Null, Member,
 class Namespace(object):
     def __init__(self):
         self.elements = [Number, String, Boolean, Null, Member, Array, Object]
+
+    def register(self, element):
+        self.elements.append(element)
+        return element