+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Refract Python Changelog
 
+## Master
+
+### Enhancements
+
+- Added [API Elements](http://api-elements.readthedocs.io) Profile.
+
+
 ## 0.1.0 (2017-04-28)
 
 Initial Release.